requirements.txt
# ============================================
# 업무 일정 관리 시스템 - 의존성 파일
# 위치: C:\Users\user\Desktop\인공지능산업협회AI\requirements.txt
# ============================================

# === 웹 프레임워크 ===
flask>=3.0.0
flask-login>=0.6.0
flask-sqlalchemy>=3.1.0
flask-bcrypt>=1.0.0
Werkzeug>=3.0.0

# === 문서 파싱 ===
python-docx>=1.1.0
pdfplumber>=0.10.0
olefile>=0.47
openpyxl>=3.1.0

# === AI (Groq API) ===
groq>=0.4.0

# === 유틸리티 ===
python-dateutil>=2.8.0
python-dotenv>=1.0.0
